# gateway-service/controller/cart_controller.go
package controller

import (
	"auth-services/model"
	"auth-services/repository"
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
)

type CartController interface {
	AddCart(*gin.Context)
	DeleteCart(*gin.Context)
	DetailCart(*gin.Context)
	ListCart(*gin.Context)
}

type cartController struct {
	allRepo repository.CartRepo
}

// CreateCart	godoc
// @Summary Create Cart
// @Description Add Product to Cart
// @param	cary body model.RequestAddCart{} true "Create Cart"
// @Produce applicaton/json
// @Success 200 {object} model.ResponseSuccessCart{}
// @Router /api/shop/add [post]
// @Tags cart-service
// AddCart implements CartController.
func (r cartController) AddCart(ctx *gin.Context) {
	var req model.RequestAddCart
	if err := ctx.ShouldBindJSON(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, "informasi salah")
		return
	}

	data, resp, err := r.allRepo.AddCart(req)
	response := &model.Response{}
	if err != nil {
		response.Success = false
		response.Status = http.StatusInternalServerError
		response.ErrorCode = "9901"
		response.RespMessage = "Failed to get response"
		response.RespData = gin.H{"error": err.Error()}
		ctx.JSON(http.StatusInternalServerError, response)
		return
	}

	if resp.StatusCode() != 200 {
		response.Success = false
		response.Status = resp.StatusCode()
		response.ErrorCode = "9901"
		response.RespMessage = "Failed to get response"
		response.RespData = gin.H{"error": fmt.Sprint(resp)}
		ctx.JSON(resp.StatusCode(), response)
		return
	}

	response.Success = true
	response.Status = http.StatusOK
	response.ErrorCode = ""
	response.RespMessage = "Success to get response"
	response.RespData = data
	ctx.JSON(http.StatusOK, response)
}

// DeleteCart	godoc
// @Summary Delete Cart
// @Description Delete Product from Cart
// @Produce applicaton/json
// @Success 200 {object} model.ResponseCart{}
// @Router /api/shop/delete [delete]
// @Param cart_id query string true "cart ID"
// @Tags cart-service
// DeleteCart implements CartController.
func (r cartController) DeleteCart(ctx *gin.Context) {
	var req model.RequesCardById
	if err := ctx.ShouldBind(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, "informasi salah")
		return
	}

	data, resp, err := r.allRepo.DeleteCart(req)
	response := &model.Response{}
	if err != nil {
		response.Success = false
		response.Status = http.StatusInternalServerError
		response.ErrorCode = "9901"
		response.RespMessage = "Failed to get response"
		response.RespData = gin.H{"error": err.Error()}
		ctx.JSON(http.StatusInternalServerError, response)
		return
	}

	if resp.StatusCode() != 200 {
		response.Success = false
		response.Status = resp.StatusCode()
		response.ErrorCode = "9901"
		response.RespMessage = "Failed to get response"
		response.RespData = gin.H{"error": fmt.Sprint(resp)}
		ctx.JSON(resp.StatusCode(), response)
		return
	}

	response.Success = true
	response.Status = http.StatusOK
	response.ErrorCode = ""
	response.RespMessage = "Success to get response"
	response.RespData = data
	ctx.JSON(http.StatusOK, response)
}

// DetailCart	godoc
// @Summary Detail Cart
// @Description Detail Product from Cart
// @Produce applicaton/json
// @Success 200 {object} model.ResponseCart{}
// @Router /api/shop/detail [get]
// @Param cart_id query string true "cart ID"
// @Tags cart-service
// DetailCart implements CartController.
func (r cartController) DetailCart(ctx *gin.Context) {
	var req model.RequesCardById
	if err := ctx.ShouldBind(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, "informasi salah")
		return
	}

	data, resp, err := r.allRepo.DetailCart(req)
	response := &model.Response{}
	if err != nil {
		response.Success = false
		response.Status = http.StatusInternalServerError
		response.ErrorCode = "9901"
		response.RespMessage = "Failed to get response"
		response.RespData = gin.H{"error": err.Error()}
		ctx.JSON(http.StatusInternalServerError, response)
		return
	}

	if resp.StatusCode() != 200 {
		response.Success = false
		response.Status = resp.StatusCode()
		response.ErrorCode = "9901"
		response.RespMessage = "Failed to get response"
		response.RespData = gin.H{"error": fmt.Sprint(resp)}
		ctx.JSON(resp.StatusCode(), response)
		return
	}

	response.Success = true
	response.Status = http.StatusOK
	response.ErrorCode = ""
	response.RespMessage = "Success to get response"
	response.RespData = data
	ctx.JSON(http.StatusOK, response)
}

// ListCart	godoc
// @Summary List Cart
// @Description List Cart
// @Produce applicaton/json
// @Success 200 {object} model.ResponseCart{}
// @Router /api/shop/list [get]
// @Tags cart-service
// ListCart implements CartController.
func (r cartController) ListCart(ctx *gin.Context) {

	data, resp, err := r.allRepo.ListCart()
	response := &model.Response{}
	if err != nil {
		response.Success = false
		response.Status = http.StatusInternalServerError
		response.ErrorCode = "9901"
		response.RespMessage = "Failed to get response"
		response.RespData = gin.H{"error": err.Error()}
		ctx.JSON(http.StatusInternalServerError, response)
		return
	}

	if resp.StatusCode() != 200 {
		response.Success = false
		response.Status = resp.StatusCode()
		response.ErrorCode = "9901"
		response.RespMessage = "Failed to get response"
		response.RespData = gin.H{"error": fmt.Sprint(resp)}
		ctx.JSON(resp.StatusCode(), response)
		return
	}

	response.Success = true
	response.Status = http.StatusOK
	response.ErrorCode = ""
	response.RespMessage = "Success to get response"
	response.RespData = data
	ctx.JSON(http.StatusOK, response)
}

func NewCartController(repo repository.CartRepo) CartController {
	return cartController{
		allRepo: repo,
	}
}