# gateway-service/docs/swagger.yaml
basePath: /api
definitions:
  model.AddCustomer:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    type: object
  model.AddOrder:
    properties:
      customer_id:
        type: integer
      order_date:
        type: string
      product_id:
        type: integer
      qty:
        type: integer
      status:
        type: string
      total_amount:
        type: number
    type: object
  model.Login:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  model.RequestAdd:
    properties:
      category:
        type: string
    type: object
  model.RequestAddCart:
    properties:
      customer_id:
        type: integer
      product_id:
        type: integer
      qty:
        type: integer
    type: object
...
  model.RequestUpdateCategory:
    properties:
      category:
        type: string
      category_id:
        type: integer
    type: object
...
  model.ResponsOrder:
    properties:
      msg:
        type: string
      order_id:
        type: integer
    type: object
...
  model.ResponseCart:
    properties:
      name:
        type: string
      product_id:
        type: integer
      qty:
        type: integer
      total_amount:
        type: number
    type: object
  model.ResponseCategory:
    properties:
      category:
        type: string
      category_id:
        type: integer
    type: object
...
  model.ResponseSuccessCart:
    properties:
      msg:
        type: string
      name:
        type: string
      price:
        type: number
      product_id:
        type: integer
      qty:
        type: integer
    type: object
...
  model.SuccessAddCustomer:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    type: object
host: 43.218.113.72:8081
info:
  contact: {}
  description: Service API in Go Using Gin framework
  title: Synapsis E-commerce API
paths:
  /api/order/add:
    post:
      description: Add Product to Order
      parameters:
      - description: Create Order
        in: body
        name: cary
        required: true
        schema:
          $ref: '#/definitions/model.AddOrder'
      produces:
      - applicaton/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponsOrder'
      summary: Create Order
      tags:
      - order-service
...
  /api/shop/add:
    post:
      description: Add Product to Cart
      parameters:
      - description: Create Cart
        in: body
        name: cary
        required: true
        schema:
          $ref: '#/definitions/model.RequestAddCart'
      produces:
      - applicaton/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccessCart'
      summary: Create Cart
      tags:
      - cart-service
  /api/shop/delete:
    delete:
      description: Delete Product from Cart
      parameters:
      - description: cart ID
        in: query
        name: cart_id
        required: true
        type: string
      produces:
      - applicaton/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseCart'
      summary: Delete Cart
      tags:
      - cart-service
  /api/shop/detail:
    get:
      description: Detail Product from Cart
      parameters:
      - description: cart ID
        in: query
        name: cart_id
        required: true
        type: string
      produces:
      - applicaton/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseCart'
      summary: Detail Cart
      tags:
      - cart-service
...
  /category/add:
    post:
      description: Create Category
      parameters:
      - description: Signin
        in: body
        name: signin
        required: true
        schema:
          $ref: '#/definitions/model.RequestAdd'
      produces:
      - applicaton/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseCategory'
      summary: Create Category
      tags:
      - category-service
  /category/detail:
    get:
      description: Detail Category
      parameters:
      - description: Category ID
        in: query
        name: category_id
        required: true
        type: string
      produces:
      - applicaton/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseCategory'
      summary: Detail Category
      tags:
      - category-service
...
  /category/update:
    post:
      description: Update Category
      parameters:
      - description: Signin
        in: body
        name: signin
        required: true
        schema:
          $ref: '#/definitions/model.RequestUpdateCategory'
      produces:
      - applicaton/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseCategory'
      summary: Update Category
      tags:
      - category-service
...
  /register/:
    post:
      description: Create User
      parameters:
      - description: Signin
        in: body
        name: signin
        required: true
        schema:
          $ref: '#/definitions/model.AddCustomer'
      produces:
      - applicaton/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessAddCustomer'
      summary: Register
      tags:
      - customer-service
  /signin/:
    post:
      description: Authentication and Unauthorized User
      parameters:
      - description: Signin
        in: body
        name: signin
        required: true
        schema:
          $ref: '#/definitions/model.Login'
      produces:
      - applicaton/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Login'
      summary: Signin
      tags:
      - customer-service
swagger: "2.0"